pyproject.toml
[project]
name = "chezzibot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.8.0",
    "coloredlogs>=15.0.1",
    "discord-ext-menus",
    "discord-py",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "pynacl>=1.5.0",
    "python-dotenv>=1.0.0",
]

[tool.uv.sources]
discord-py = { git = "https://github.com/Rapptz/discord.py" }
discord-ext-menus = { git = "https://github.com/Rapptz/discord-ext-menus" }

[tool.pyright]
venvPath = "/home/chezzijr/proj/chezzibot"
venv = ".venv"
